refactor(app): share duplicated types and helpers
Hunt down and consolidate repeated TypeScript across the app, server and
shared layers.

Types:
- SourceInfoCard/CiEnvCard/TestCaseSummary drop their inline ScmInfo/CiInfo
  and reuse the canonical TestRunScmMetadata/TestRunCiMetadata from types/api
- ConsoleEntry, NetworkRequest, ServerLogEntry, EndpointSummary,
  DiagnosisCompact and PatchValidation local copies now import their existing
  canonical definitions
- ContextSection (4 copies) is defined once in types/api; the diagnosis
  composable re-exports it for its consumers
- The DbClient/DB handle, re-derived ~22x across server/utils, is exported
  once from server/database and imported everywhere

Functions:
- sha256Hex -> shared/utils/hash
- buildCompareUrl/computeMetadataDiff/getBrowserList -> shared/utils/run-metadata
- sanitizeFilename -> server/utils/sanitize-filename
- asCategory is exported from shared/ai-diagnosis and reused by ai-research
- the two inline isImage copies use the existing isImageFile util
- passRate/getPassRate/getTestCaseStatus/BadgeColor -> app/utils

Genuinely divergent same-name shapes (Subscription/Channel, DomSnapshotResponse,
the WebVitals read-lens, RecordedReq test mocks) and different-behavior helpers
(escapeHtml/escapeText/stripAnsi) are intentionally left alone.
