+++ application/server/utils/ai-diagnosis.ts
@@ -22,8 +22,7 @@ import { nameNewClusters } from './cluster-naming';
 import { RESEARCH_SYSTEM_PROMPT, RESEARCH_JSON_SCHEMA, parseResearchJson, formatResearchBlock } from './ai-research';
 import { buildDiagnosisVersionValues } from '#shared/handlers/diagnosis-versions';
 import { emitNotification } from './notifications/emit';
-
-type DbClient = Awaited<ReturnType<typeof import('../database').getDatabase>>;
+import type { DbClient } from '../database';
 
 const STALE_RUNNING_MS = 5 * 60 * 1000;
