+++ application/server/utils/cancel-instance-runs.ts
@@ -1,8 +1,6 @@
 import { runEventBus } from './run-events';
 import { cancelInstanceRuns as sharedCancelInstanceRuns } from '#shared/handlers/failure-cluster-ops';
-import type { getDatabase } from '../database';
-
-type DB = Awaited<ReturnType<typeof getDatabase>>;
+import type { DbClient as DB } from '../database';
 
 export async function cancelInstanceRuns(
   db: DB,
