+++ application/server/utils/persist-run-cases.ts
@@ -9,9 +9,7 @@ import { SUITE_PATH_SEP, joinSuitePath } from '#shared/utils/suites';
 import { getOrCreateFailureClusters, type PendingCluster } from '#shared/handlers/failure-cluster-ops';
 import { upsertLocatorSnapshots } from './locator-healing';
 import type { LocatorSnapshot } from '#shared/locator-healing.types';
-import type { getDatabase } from '../database';
-
-type DB = Awaited<ReturnType<typeof getDatabase>>;
+import type { DbClient as DB } from '../database';
 
 /**
  * Normalised test-case data ready to be persisted for a run. `filePath` + `suitePath` + `title`
