+++ application/server/utils/test-case-cache.ts
@@ -1,8 +1,6 @@
 import { testCases } from '../database/schema';
 import { eq } from 'drizzle-orm';
-import type { getDatabase } from '../database';
-
-type DB = Awaited<ReturnType<typeof getDatabase>>;
+import type { DbClient as DB } from '../database';
 
 // Cache key: filePath + NUL + suitePath (raw stored text, \x1f-delimited) + NUL + title
 // NUL separates the three segments; \x1f separates describe levels within suitePath.
